Move analysis reducers to root reducers file
